backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pytesseract==0.3.13
Pillow==10.4.0
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
easyocr==1.7.2
httpx==0.27.0
pymongo==4.8.0
motor==3.5.1
certifi==2024.8.30
